+++ src/components/SubmitButton/SubmitButton.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import styles from "./SubmitButton.module.css";
 
 const SubmitButton = () => {
