+++ src/pages/MainPage/MainPage.module.css
@@ -1,0 +1,7 @@
+.app {
+  margin-top: 100px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 45px;
+}
