+++ src/stories/CardPreview.stories.tsx
@@ -13,13 +13,16 @@ type Story = StoryObj<typeof meta>;
 export const Default: Story = {
   args: {
     cardNumbers: {
-      first: "1234",
-      second: "1234",
-      third: "1234",
-      fourth: "1234",
+      first: "",
+      second: "",
+      third: "",
+      fourth: "",
     },
-    expirationDate: { month: "08", year: "31" },
-    ownerName: { ownerName: "kim" },
+    expirationDate: { month: "", year: "" },
+    ownerName: "",
+    cardCompany: "",
+    CVC: "",
+    face: "front",
   },
 };
 
@@ -32,19 +35,57 @@ export const Visa: Story = {
       fourth: "1234",
     },
     expirationDate: { month: "08", year: "31" },
-    ownerName: { ownerName: "kim" },
+    ownerName: "kim",
+    cardCompany: "",
+    CVC: "",
+    face: "front",
   },
 };
 
 export const Master: Story = {
   args: {
     cardNumbers: {
-      first: "5134",
+      first: "5434",
       second: "1234",
       third: "1234",
       fourth: "1234",
     },
     expirationDate: { month: "08", year: "31" },
-    ownerName: { ownerName: "kim" },
+    ownerName: "kim",
+    cardCompany: "",
+    CVC: "",
+    face: "front",
+  },
+};
+
+export const KAKAOBANK: Story = {
+  args: {
+    cardNumbers: {
+      first: "5434",
+      second: "1234",
+      third: "1234",
+      fourth: "1234",
+    },
+    expirationDate: { month: "08", year: "31" },
+    ownerName: "kim",
+    cardCompany: "카카오뱅크",
+    CVC: "",
+    face: "front",
+  },
+};
+
+export const BACK: Story = {
+  args: {
+    cardNumbers: {
+      first: "5434",
+      second: "1234",
+      third: "1234",
+      fourth: "1234",
+    },
+    expirationDate: { month: "08", year: "31" },
+    ownerName: "kim",
+    cardCompany: "카카오뱅크",
+    CVC: "",
+    face: "back",
   },
 };
